+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add Empty static results in KeysetPaginationResult and OffsetPaginationResult.
 - Move `PaginationActionDetector` to main package.
+- Add `ConfigurePagination` for swagger and make `AddPaginationOperationFilter` obsolete.